requirements.txt
agno>=1.0.0
openai>=1.30.0
serpapi>=0.1.5
google-search-results>=2.4.2
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pyswip>=0.2.12
python-docx>=1.1.0
redis>=5.0.0
